scripts/test_requirements.txt
# Additional requirements for testing
requests==2.31.0
python-docx==1.1.0
